REST API Ver 1

+++ tourdest/views.py
@@ -13,11 +13,18 @@ from tourdest.models import Product
 from tourdest.models import Payment
 from tourdest.models import PaymentDetail
 from tourdest.serializers import UserSerializer
+from tourdest.serializers import UserLoginSerializer
 from tourdest.serializers import ShopSerializer
 from tourdest.serializers import ShopPositionSerializer
 from tourdest.serializers import ProductSerializer
 from tourdest.serializers import PaymentSerializer
 from tourdest.serializers import PaymentDetailSerializer
+from rest_framework.authtoken.models import Token
+from django.contrib.auth import authenticate
+from rest_framework.decorators import api_view, authentication_classes, permission_classes
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.authentication import TokenAuthentication
+from django.contrib.auth import logout
 
 class JSONResponse(HttpResponse):
     def __init__(self, data, **kwargs):
@@ -26,8 +33,13 @@ class JSONResponse(HttpResponse):
         super(JSONResponse, self).__init__(content, **kwargs)
 
 @csrf_exempt
+@api_view(['GET'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def user_list(request, lev = "all", stat = "all"):
     if request.method == 'GET':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         if stat == "all":
             if lev == "all":
                 users = User.objects.all()
@@ -56,15 +68,47 @@ def user_list(request, lev = "all", stat = "all"):
                 users_serializer = UserSerializer(users, many=True)
                 return JSONResponse(users_serializer.data)
 
-    elif request.method == 'POST':
+@csrf_exempt
+def user_create(request):
+    if request.method == 'POST':
         user_data = JSONParser().parse(request)
         user_serializer = UserSerializer(data=user_data)
         if user_serializer.is_valid():
-            user_serializer.save()
-            return JSONResponse(user_serializer.data, status=status.HTTP_201_CREATED)
+            user = user_serializer.save()
+            token, created = Token.objects.get_or_create(user=user)
+            return JSONResponse({'data':user_serializer.data, 'token': token.key}, status=status.HTTP_201_CREATED)
         return JSONResponse(user_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @csrf_exempt
+def user_login(request):
+    if request.method == 'POST':
+        user_data = JSONParser().parse(request)
+        user_serializer = UserLoginSerializer(data=user_data)
+        if user_serializer.is_valid():
+            user = authenticate(request=request, username=user_data['username'], password=user_data['password'])
+            if user is not None:
+                token, _ = Token.objects.get_or_create(user=user)
+                return JSONResponse({'data': UserSerializer(user).data, 'token': token.key})
+            else:
+                return JSONResponse(user_serializer.errors, status=status.HTTP_401_UNAUTHORIZED)
+        return JSONResponse(user_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+@csrf_exempt
+@api_view(['POST'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
+def user_logout(request):
+    if request.method == "POST":
+        request.user.auth_token.delete()
+
+        logout(request)
+
+        return JSONResponse("User Logout Successfully", status=status.HTTP_200_OK)
+
+@csrf_exempt
+@api_view(['GET', 'PUT', 'PATCH', 'DELETE'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def user_detail(request, pk):
     try:
         user = User.objects.get(pk=pk)
@@ -92,10 +136,15 @@ def user_detail(request, pk):
         return JSONResponse(user_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'DELETE':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         user.delete()
         return HttpResponse(status=status.HTTP_204_NO_CONTENT)
 
 @csrf_exempt
+@api_view(['GET', 'POST'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def shop_list(request, stat = "all"):
     if request.method == 'GET':
         if stat == "all":
@@ -112,6 +161,8 @@ def shop_list(request, stat = "all"):
             return JSONResponse(shops_serializer.data)
 
     elif request.method == 'POST':
+        if request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         shop_data = JSONParser().parse(request)
         shop_serializer = ShopSerializer(data=shop_data)
         if shop_serializer.is_valid():
@@ -120,6 +171,9 @@ def shop_list(request, stat = "all"):
         return JSONResponse(shop_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @csrf_exempt
+@api_view(['GET', 'PUT', 'PATCH', 'DELETE'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def shop_detail(request, pk, stat = "all"):
     try:
         if stat == "all":
@@ -136,6 +190,8 @@ def shop_detail(request, pk, stat = "all"):
         return JSONResponse(shop_serializer.data)
 
     elif request.method == 'PUT':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         shop_data = JSONParser().parse(request)
         shop_serializer = ShopSerializer(shop, data=shop_data)
         if shop_serializer.is_valid():
@@ -144,6 +200,8 @@ def shop_detail(request, pk, stat = "all"):
         return JSONResponse(shop_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'PATCH':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         shop_data = JSONParser().parse(request)
         shop_serializer = ShopSerializer(shop, data=shop_data, partial=True)
         if shop_serializer.is_valid():
@@ -152,17 +210,26 @@ def shop_detail(request, pk, stat = "all"):
         return JSONResponse(shop_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'DELETE':
+        if request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         shop.delete()
         return HttpResponse(status=status.HTTP_204_NO_CONTENT)
 
 @csrf_exempt
+@api_view(['GET', 'POST'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def shoppos_list(request):
     if request.method == 'GET':
-            shoppos = ShopPosition.objects.all()
-            shoppos_serializer = ShopPositionSerializer(shoppos, many=True)
-            return JSONResponse(shoppos_serializer.data)
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
+        shoppos = ShopPosition.objects.all()
+        shoppos_serializer = ShopPositionSerializer(shoppos, many=True)
+        return JSONResponse(shoppos_serializer.data)
 
     elif request.method == 'POST':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         shoppos_data = JSONParser().parse(request)
         shoppos_serializer = ShopPositionSerializer(data=shoppos_data)
         if shoppos_serializer.is_valid():
@@ -171,13 +238,21 @@ def shoppos_list(request):
         return JSONResponse(shoppos_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @csrf_exempt
+@api_view(['GET'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def shoppos_get(request, shop):
     if request.method == 'GET':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         shoppos = ShopPosition.objects.all().filter(shop_id = shop)
         shoppos_serializer = ShopPositionSerializer(shoppos, many=True)
         return JSONResponse(shoppos_serializer.data)
 
 @csrf_exempt
+@api_view(['GET', 'PUT', 'PATCH', 'DELETE'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def shoppos_detail(request, pk):
     try:
         shoppos = ShopPosition.objects.get(pk=pk)
@@ -185,10 +260,14 @@ def shoppos_detail(request, pk):
         return HttpResponse(status=status.HTTP_404_NOT_FOUND)
 
     if request.method == 'GET':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         shoppos_serializer = ShopPositionSerializer(shoppos)
         return JSONResponse(shoppos_serializer.data)
 
     elif request.method == 'PUT':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         shoppos_data = JSONParser().parse(request)
         shoppos_serializer = ShopPositionSerializer(shoppos, data=shoppos_data)
         if shoppos_serializer.is_valid():
@@ -197,6 +276,8 @@ def shoppos_detail(request, pk):
         return JSONResponse(shoppos_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'PATCH':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         shoppos_data = JSONParser().parse(request)
         shoppos_serializer = ShopPositionSerializer(shoppos, data=shoppos_data, partial = True)
         if shoppos_serializer.is_valid():
@@ -205,10 +286,15 @@ def shoppos_detail(request, pk):
         return JSONResponse(shoppos_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'DELETE':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         shoppos.delete()
         return HttpResponse(status=status.HTTP_204_NO_CONTENT)
 
 @csrf_exempt
+@api_view(['GET', 'POST'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def product_list(request, stat = "all", fk = 0):
     if request.method == 'GET':
         if stat == "all":
@@ -240,6 +326,8 @@ def product_list(request, stat = "all", fk = 0):
                 return JSONResponse(products_serializer.data)
 
     elif request.method == 'POST':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         product_data = JSONParser().parse(request)
         product_serializer = ProductSerializer(data=product_data)
         if product_serializer.is_valid():
@@ -248,6 +336,9 @@ def product_list(request, stat = "all", fk = 0):
         return JSONResponse(product_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @csrf_exempt
+@api_view(['GET', 'PUT', 'PATCH', 'DELETE'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def product_detail(request, pk, stat = "all"):
     try:
         if stat == "all":
@@ -264,6 +355,8 @@ def product_detail(request, pk, stat = "all"):
         return JSONResponse(product_serializer.data)
 
     elif request.method == 'PUT':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         product_data = JSONParser().parse(request)
         product_serializer = ProductSerializer(product, data=product_data)
         if product_serializer.is_valid():
@@ -272,6 +365,8 @@ def product_detail(request, pk, stat = "all"):
         return JSONResponse(product_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'PATCH':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         product_data = JSONParser().parse(request)
         product_serializer = ProductSerializer(product, data=product_data, partial = True)
         if product_serializer.is_valid():
@@ -280,12 +375,19 @@ def product_detail(request, pk, stat = "all"):
         return JSONResponse(product_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'DELETE':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         product.delete()
         return HttpResponse(status=status.HTTP_204_NO_CONTENT)
 
 @csrf_exempt
+@api_view(['GET', 'POST'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def payment_list(request, stat = "all"):
     if request.method == 'GET':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         if stat == "all":
             payments = Payment.objects.all()
             payments_serializer = PaymentSerializer(payments, many=True)
@@ -308,6 +410,9 @@ def payment_list(request, stat = "all"):
         return JSONResponse(payment_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @csrf_exempt
+@api_view(['GET', 'PUT', 'PATCH', 'DELETE'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def payment_detail(request, pk, stat = "all"):
     try:
         if stat == "all":
@@ -324,6 +429,8 @@ def payment_detail(request, pk, stat = "all"):
         return JSONResponse(payment_serializer.data)
 
     elif request.method == 'PUT':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         payment_data = JSONParser().parse(request)
         payment_serializer = PaymentSerializer(payment, data=payment_data)
         if payment_serializer.is_valid():
@@ -332,6 +439,8 @@ def payment_detail(request, pk, stat = "all"):
         return JSONResponse(payment_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'PATCH':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         payment_data = JSONParser().parse(request)
         payment_serializer = PaymentSerializer(payment, data=payment_data, partial = True)
         if payment_serializer.is_valid():
@@ -340,10 +449,15 @@ def payment_detail(request, pk, stat = "all"):
         return JSONResponse(payment_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'DELETE':
+        if request.user.level != "A" and request.user.level != "SA":
+            return JSONResponse({'detail': 'You do not have permission to access this resource.'}, status=status.HTTP_403_FORBIDDEN)
         payment.delete()
         return HttpResponse(status=status.HTTP_204_NO_CONTENT)
 
 @csrf_exempt
+@api_view(['GET'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def payment_get(request, user, shop, stat = "all"):
     if request.method == 'GET':
         if stat == "all":
@@ -405,6 +519,9 @@ def payment_get(request, user, shop, stat = "all"):
                     return JSONResponse(payments_serializer.data)
 
 @csrf_exempt
+@api_view(['POST'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def paymentdet_list(request):
     if request.method == 'POST':
         paymentdet_data = JSONParser().parse(request)
@@ -415,6 +532,9 @@ def paymentdet_list(request):
         return JSONResponse(paymentdet_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @csrf_exempt
+@api_view(['GET'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def paymentdet_get(request, payment, product):
     if request.method == 'GET':
         if payment == "0":
@@ -437,6 +557,9 @@ def paymentdet_get(request, payment, product):
                 return JSONResponse(paymentdets_serializer.data)
 
 @csrf_exempt
+@api_view(['GET', 'PUT', 'PATCH', 'DELETE'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def paymentdet_detail(request, pk):
     try:
         paymentdet = PaymentDetail.objects.get(pk=pk)
